Design: 대신 정해주기 레이아웃 변경
