+++ app/src/App.jsx
@@ -4,7 +4,7 @@ import "./App.css";
 import Home from "./routes/Home";
 import Login from "./routes/Login";
 import MyPage from "./routes/MyPage";
-import PlaySelect from "./routes/PlaySelect";
+import PlaySelect from "./routes/PlaySelectpage/PlaySelect";
 import PlayWorldcup from "./routes/PlayWorldcup";
 import RegisterSelect from "./routes/RegisterSelect";
 import RegisterWorldcup from "./routes/RegisterWorldcup";
